Refactor the middleware into separate files
